Compiles and runs but crashes

+++ libprojectM210Regal/Renderer/VideoEcho.cpp
@@ -11,8 +11,9 @@
 #include <GLES/gl.h>
 #else
 #ifdef __APPLE__
-#include <OpenGL/gl.h>
-#include <OpenGL/glu.h>
+//#include <OpenGL/gl.h>
+//#include <OpenGL/glu.h>
+#include "Regal.h"
 #else
 #include <GL/gl.h>
 #include <GL/glu.h>
@@ -77,10 +78,10 @@ void VideoEcho::Draw(RenderContext &context)
 			default: flipx=1;flipy=1; break;
 		}
 
-		float pointsFlip[4][2] = {{-0.5*flipx, -0.5*flipy},
-					  {-0.5*flipx,  0.5*flipy},
-					  { 0.5*flipx,  0.5*flipy},
-					  { 0.5*flipx, -0.5*flipy}};
+		float pointsFlip[4][2] = {{static_cast<float>(-0.5*flipx), static_cast<float>(-0.5*flipy)},
+					  {static_cast<float>(-0.5*flipx),  static_cast<float>(0.5*flipy)},
+					  { static_cast<float>(0.5*flipx),  static_cast<float>(0.5*flipy)},
+					  { static_cast<float>(0.5*flipx), static_cast<float>(-0.5*flipy)}};
 
 		glVertexPointer(2,GL_FLOAT,0,pointsFlip);
 		glDrawArrays(GL_TRIANGLE_FAN,0,4);
